add improved CookieHelper class
